requirements.txt
arch==5.3.1
darts>=0.27.0
matplotlib>=3.7.0
neuralprophet>=0.6.0
numpy>=1.24.0
pandas>=1.5.0
prophet>=1.1.5
pytorch-forecasting>=1.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.31.0
scipy>=1.10.0
torch>=2.0.0
tqdm>=4.65.0
